+++ example/example.js
@@ -1,19 +1,10 @@
 import { instantiate, invoke } from "./__dart/example.mjs";
 import dartModule from "./__dart/example.wasm";
+import { exec } from "../web/exec.mjs";
 
 export default {
   async fetch(request, _env, _ctx) {
-    const dartInstance = await instantiate(
-      dartModule,
-    );
-
-    let response;
-    globalThis.__dart_cf_workers = {
-      request: () => request,
-      response: (r) => response = r,
-    };
-    await invoke(dartInstance);
-    console.log(response);
-    return response;
+    const dartInstance = await instantiate(dartModule);
+    return exec(() => invoke(dartInstance), request);
   },
 };